requirements/dev-requirements.txt
-r test-requirements.txt
autopep8==1.2.2
django-debug-toolbar==1.4
ipdb==0.9.0
pytest-ipdb==0.1-prerelease2
